wip on take price

+++ contracts/vm/ops/core/OpLoopN.sol
@@ -19,11 +19,11 @@ library OpLoopN {
         StackTop stackTop_
     ) internal view returns (StackTop) {
         unchecked {
-            uint n_ = Operand.unwrap(operand_) & 0x0F;
+            uint256 n_ = Operand.unwrap(operand_) & 0x0F;
             SourceIndex loopSourceIndex_ = SourceIndex.wrap(
                 (Operand.unwrap(operand_) & 0xF0) >> 4
             );
-            for (uint i_ = 0; i_ <= n_; i_++) {
+            for (uint256 i_ = 0; i_ <= n_; i_++) {
                 stackTop_ = integrityState_.ensureIntegrity(
                     integrityState_,
                     loopSourceIndex_,
